% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{bayesOptCV}
\alias{bayesOptCV}
\title{Bayesian optimization with cross validation}
\usage{
bayesOptCV(
  rec = NULL,
  model = NULL,
  v = "5",
  trainingData = NULL,
  gridNum = NULL,
  iter = 10,
  seed = "4814"
)
}
\arguments{
\item{rec}{데이터, 전처리 정보를 포함한 recipe object}

\item{model}{hyperparameters, ngine, mode 정보가 포함된 model object}

\item{v}{v-fold cross validation을 진행 (default: 5, 각 fold 별로 30개 이상의 observations가 있어야 유효한 모델링 결과를 얻을 수 있습니다.)}

\item{trainingData}{훈련데이터 셋}

\item{iter}{grid search를 수행할 때 각 hyperparameter의 값을 담은 object}

\item{seed}{seed값 설정}

\item{initial}{몇 개의 grid로}
}
\description{
Bayesian optimization with cross validation
}
\details{
교차검증 수행 과정에서, Bayesian optimization을 통해 모델의 하이퍼파라미터를 최적화합니다.
}
