% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{clusteringVis}
\alias{clusteringVis}
\title{clusteringVis}
\usage{
clusteringVis(data = NULL, model = NULL, nStart = 25, maxK = 10)
}
\arguments{
\item{data}{data}

\item{model}{model}

\item{nStart}{nStart}

\item{maxK}{maxK}
}
\description{
clusteringVis
}
\details{
clusteringVis
}
