% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Confusion matrix}
\usage{
confusionMatrix(i, models_list, targetVar)
}
\arguments{
\item{i}{i}

\item{models_list}{models_list}

\item{targetVar}{targetVar}
}
\description{
Confusion matrix
}
\details{
Confusion matrix
}
