% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{evalMetrics}
\alias{evalMetrics}
\title{Evaluation metrics}
\usage{
evalMetrics(models_list, targetVar)
}
\arguments{
\item{models_list}{models_list}

\item{targetVar}{targetVar}

\item{i}{i}
}
\description{
Evaluation metrics
}
\details{
Evaluation metrics
}
