% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{evalMetricsR}
\alias{evalMetricsR}
\title{Evaluation metrics for Regression}
\usage{
evalMetricsR(modelsList, targetVar)
}
\arguments{
\item{modelsList}{ML 모델 리스트}

\item{targetVar}{타겟 변수}
}
\description{
Evaluation metrics for Regression
}
\details{
ML 모델 리스트로부터 Regression 모델들에 대한 Evaluation metrics를 생성합니다.
}
