% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitBestModel}
\alias{fitBestModel}
\title{fitting in best model}
\usage{
fitBestModel(
  gridSearchResult,
  metric,
  model,
  formula,
  trainingData,
  splitedData,
  algo
)
}
\arguments{
\item{gridSearchResult}{gridSearchCV의 결과값}

\item{metric}{모델의 성능을 평가할 기준지표 (classification: "roc_auc" (default), "accuracy" / regression: "rmse" (default), "rsq")}

\item{model}{hyperparameters, ngine, mode 정보가 포함된 model object}

\item{formula}{모델링을 위한 수식}

\item{trainingData}{훈련데이터 셋}

\item{splitedData}{train-test 데이터 분할 정보를 포함하고 있는 전체 데이터 셋}

\item{algo}{사용자가 임의로 지정할 알고리즘명 (default: "linear Regression")}
}
\description{
fitting in best model
}
\details{
fitting in best model // tune workflows
}
