% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{gridSerachCV}
\alias{gridSerachCV}
\title{Grid Search with cross validation}
\usage{
gridSerachCV(rec, model, v = 5, data, parameterGrid)
}
\arguments{
\item{rec}{rec}

\item{model}{model}

\item{v}{v-fold CV}

\item{data}{data}

\item{parameter_grid}{parameter_grid}
}
\description{
Grid Search with cross validation
}
\details{
Grid Search with cross validation
}
