% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{kMeansClustering}
\alias{kMeansClustering}
\title{K means clustering}
\usage{
kMeansClustering(
  data,
  maxK = 15,
  nStart = 25,
  iterMax = 10,
  nBoot = 100,
  algorithm = "Hartigan-Wong",
  selectOptimal = "silhouette",
  seedNum = 6471
)
}
\arguments{
\item{data}{전처리가 완료된 데이터}

\item{maxK}{클러스터링 수행 시 군집을 2, 3, ..., maxK개로 분할 (default: 15)}

\item{iterMax}{반복계산을 수행할 최대 횟수 (default: 10)}

\item{nBoot}{gap statictic을 사용해 클러스터링을 수행할 때 Monte Carlo (bootstrap) 샘플의 개수 (selectOptimal == "gap_stat" 일 경우에만 지정, default: 100)}

\item{algorithm}{K means를 수행할 알고리즘 선택 ("Hartigan-Wong" (default), "Lloyd", "Forgy", "MacQueen")}

\item{selectOptimal}{최적의 K값을 선정할 때 사용할 method 선택 ("silhouette" (default), "gap_stat")}

\item{seedNum}{seed값 설정}

\item{nstart}{랜덤 샘플에 대해 초기 클러스터링을 nstart번 시행 (default: 25)}
}
\description{
K means clustering
}
\details{
K means clustering
selectOptimal: silhouette, gap_stat
hyperparameters: maxK, nstart
}
