% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kMeansClustering_phi}
\alias{kMeansClustering_phi}
\title{K means clustering}
\usage{
kMeansClustering_phi(
  data,
  maxK = 10,
  nstart = 25,
  selectOptimal = "silhouette",
  seed_num = 6471
)
}
\arguments{
\item{data}{data}

\item{maxK}{maxK}

\item{nstart}{nstart}

\item{selectOptimal}{selectOptimal}

\item{seed_num}{seed_num}
}
\description{
K means clustering
}
\details{
K means clustering
selectOptimal: silhouette, gap_stat
hyperparameters: maxK, nstart
}
