% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{KNN}
\alias{KNN}
\title{K-Nearest Neighbors}
\usage{
KNN(
  algo = "KNN",
  engine = "kknn",
  mode = "classification",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  neighborsRangeMin = 1,
  neighborsRangeMax = 10,
  neighborsRangeLevels = 10,
  metric = NULL
)
}
\arguments{
\item{algo}{사용자가 임의로 지정할 알고리즘명 (default: "KNN")}

\item{engine}{모델을 생성할 때 사용할 패키지 ("kknn" (default))}

\item{mode}{분석 유형 ("classification" (default), "regression")}

\item{trainingData}{훈련데이터 셋}

\item{splitedData}{train-test 데이터 분할 정보를 포함하고 있는 전체 데이터 셋}

\item{formula}{모델링을 위한 수식}

\item{rec}{데이터, 전처리 정보를 포함한 recipe object}

\item{v}{v-fold cross validation을 진행 (default: 5, 각 fold 별로 30개 이상의 observations가 있어야 유효한 모델링 결과를 얻을 수 있습니다.)}

\item{metric}{모델의 성능을 평가할 기준지표 (classification: "roc_auc" (default), "accuracy" / regression: "rmse" (default), "rsq")}

\item{...}{hyperparameters의 범위에 대한 Min, Max, Levels 값에 해당하는 파라미터를 지정합니다.}
}
\description{
K-Nearest Neighbors
}
\details{
KNN 알고리즘 함수.
데이터로부터 거리가 가까운 K개의 다른 데이터의 레이블을 참조하여 분류하는 알고리즘
hyperparameters: neighbors
}
