% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{linearRegression_phi}
\alias{linearRegression_phi}
\title{Linear Regression}
\usage{
linearRegression_phi(engine = "lm", mode = "regression")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
Linear Regression
}
\details{
Linear Regression
hyperparameters: penalty, mixture
}
