% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervisedLearning.R
\name{logisticRegression_phi}
\alias{logisticRegression_phi}
\title{Logistic Regression}
\usage{
logisticRegression_phi(
  penalty = NULL,
  mixture = NULL,
  engine = "glm",
  mode = "classification"
)
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}

\item{data}{분석에 사용할 변수들이 포함된 데이터}
}
\description{
Logistic Regression
}
\details{
로지스틱 회귀 알고리즘 함수. 예측 변수들이 정규분포를 따르지 않아도 사용할 수 있습니다.
그러나 이 알고리즘은 결과 변수가 선형적으로 구분되며, 예측 변수들의 값이 결과 변수와 선형 관계를
가진다고 가정합니다. 만약 데이터가 이 가정을 충족하지 않는 경우 성능이 저하될 수 있습니다.
}
