% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{mlp_phi}
\alias{mlp_phi}
\title{neural network}
\usage{
mlp_phi(engine = "nnet", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
neural network
}
\details{
neural network 알고리즘 함수.
neural network 모델은 생물학적인 뉴런을 수학적으로 모델링한 것.
여러개의 뉴런으로부터 입력값을 받아서 세포체에 저장하다가 자신의 용량을 넘어서면 외부로 출력값을 내보내는 것처럼,
인공신경망 뉴런은 여러 입력값을 받아서 일정 수준이 넘어서면 활성화되어 출력값을 내보낸다.
hyperparameters: hidden_units, penalty, dropout, epochs, activation, learn_rate
}
