% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{multinomialRegression}
\alias{multinomialRegression}
\title{Multinomial Regression}
\usage{
multinomialRegression(
  algo = "logistic Regression",
  engine = "glmnet",
  mode = "classification",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  gridNum = 5,
  iter = 10,
  metric = "roc_auc",
  seed = 1234
)
}
\arguments{
\item{algo}{A name of the algorithm which can be customized by user. (default: "logistic Regression")}

\item{engine}{The name of software that should be used to fit the model. (Option: "glmnet" (default))}

\item{mode}{The model type. It should be "classification" or "regression". (Option: "classification" (default))}

\item{trainingData}{A data frame for training}

\item{splitedData}{A data frame including metadata of split}

\item{formula}{formula for modeling}

\item{rec}{Recipe object containing preprocessing information for cross-validation}

\item{v}{Applying v-fold cross validation in modeling process (default: 5)}

\item{metric}{Metric to evaluate the performance (Option: "roc_auc" (default), "accuracy")}
}
\description{
Multinomial regression is a algorithm for multinomial classification problems

This function supports: binary classification
}
\details{
Hyperparameters for tuning: penalty, mixture
}
