% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{naiveBayes}
\alias{naiveBayes}
\title{Naive Bayes}
\usage{
naiveBayes(
  algo = "Naive Bayes",
  engine = "klaR",
  mode = "classification",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  smoothnessRangeMin = 0.5,
  smoothnessRangeMax = 1.5,
  smoothnessRangeLevels = 3,
  LaplaceRangeMin = 0,
  LaplaceRangeMax = 3,
  LaplaceRangeLevels = 4,
  metric = NULL
)
}
\arguments{
\item{algo}{사용자가 임의로 지정할 알고리즘명 (default: "Naive Bayes")}

\item{engine}{모델을 생성할 때 사용할 패키지 ("klaR" (default), naivebayes)}

\item{mode}{분석 유형 ("classification" (default))}

\item{trainingData}{훈련데이터 셋}

\item{splitedData}{train-test 데이터 분할 정보를 포함하고 있는 전체 데이터 셋}

\item{formula}{모델링을 위한 수식}

\item{rec}{데이터, 전처리 정보를 포함한 recipe object}

\item{v}{v-fold cross validation을 진행 (default: 5, 각 fold 별로 30개 이상의 observations가 있어야 유효한 모델링 결과를 얻을 수 있습니다.)}

\item{metric}{모델의 성능을 평가할 기준지표 (classification: "roc_auc" (default), "accuracy" / regression: "rmse" (default), "rsq")}

\item{...}{hyperparameters의 범위에 대한 Min, Max, Levels 값에 해당하는 파라미터를 지정합니다.}
}
\description{
Naive Bayes
}
\details{
Naive Bayes
hyperparameters: smoothness, Laplace
}
