% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{naiveBayes_phi}
\alias{naiveBayes_phi}
\title{Naive Bayes}
\usage{
naiveBayes_phi(engine = "klaR", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
Naive Bayes
}
\details{
Naive Bayes
hyperparameters: smoothness, Laplace
}
