% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepForCV}
\alias{prepForCV}
\title{Preprocessing for cross validation}
\usage{
prepForCV(
  data = NULL,
  formula = NULL,
  imputation = FALSE,
  normalization = FALSE,
  nominalImputationType = "mode",
  numericImputationType = "mean",
  normalizationType = "range",
  seed = "4814"
)
}
\arguments{
\item{data}{Training dataset to apply local preprocessing recipe.}

\item{formula}{formula for modeling}

\item{imputation}{If "imputation = TRUE", the model will be trained using cross-validation with imputation.}

\item{normalization}{If "normalization = TRUE", the model will be trained using cross-validation with normalization}

\item{nominalImputationType}{Imputation method for nominal variable (Option: mode(default), bag, knn)}

\item{numericImputationType}{Imputation method for numeric variable (Option: mean(default), bag, knn, linear, lower, median, roll)}

\item{normalizationType}{Normalization method (Option: range(default), center, normalization, scale)}

\item{seed}{seed}
}
\description{
Preprocessing for cross validation
}
\details{
Define the local preprocessing method to be applied to the training data for each fold when the training data is divided into several folds.
}
