% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocessing}
\usage{
preprocessing(
  data,
  formula,
  imputationType = "mean",
  normalizationType = "range",
  pcaThres = "0.7",
  imputation = TRUE,
  normalization = TRUE,
  pca = TRUE
)
}
\arguments{
\item{data}{data}

\item{formula}{formula}

\item{imputationType}{imputationType}

\item{normalizationType}{normalizationType}

\item{imputation}{imputation}

\item{normalization}{normalization}

\item{pca}{pca}

\item{pca_thres}{pca_thres}
}
\description{
Preprocessing
}
\details{
Preprocessing
}
