% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{randomForest}
\alias{randomForest}
\title{Random Forest}
\usage{
randomForest(
  algo = "Random Forest",
  engine = "ranger",
  mode = "classification",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  gridNum = 5,
  iter = 10,
  metric = NULL,
  seed = 1234
)
}
\arguments{
\item{algo}{A name of the algorithm which can be customized by user (default: "Random Forest").}

\item{engine}{The name of software that should be used to fit the model ("rpart" (default), "randomForest", "partykit").}

\item{mode}{The model type. It should be "classification" or "regression" ("classification" (default), "regression").}

\item{trainingData}{The training data.}

\item{splitedData}{A data frame including metadata of split.}

\item{formula}{formula for modeling}

\item{rec}{Recipe object containing preprocessing information for cross-validation.}

\item{v}{Applying v-fold cross validation in modeling process (default: 5).}

\item{gridNum}{Initial number of iterations to run before starting the optimization algorithm.}

\item{iter}{The maximum number of search iterations.}

\item{metric}{Metric to evaluate the performance (classification: "roc_auc" (default), "accuracy" / regression: "rmse" (default), "rsq").}

\item{seed}{Seed for reproducible results.}
}
\description{
Random Forest
}
\details{
The function for training user-defined Random Forest model.

Hyperparameters for tuning: trees, min_n, mtry
}
