% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{randomForest_phi}
\alias{randomForest_phi}
\title{Random Forest}
\usage{
randomForest_phi(engine = "randomForest", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
Random Forest
}
\details{
랜덤 포레스트 알고리즘 함수. 여러개의 Decision Tree를 형성.
새로운 데이터 포인트를 각 트리에 동시에 통과 시켜 각 트리가 분류한 결과에서 투표를 실시하여
가장 많이 득표한 결과를 최종 분류 결과로 선택
hyperparameters: trees, min_n, mtry
}
