% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{rocCurve}
\alias{rocCurve}
\title{AUC-ROC Curve}
\usage{
rocCurve(models_list, targetVar)
}
\arguments{
\item{models_list}{models_list}

\item{targetVar}{targetVar}
}
\description{
AUC-ROC Curve
}
\details{
AUC-ROC Curve
}
