% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{SVM}
\alias{SVM}
\title{SVM}
\usage{
SVM(
  algo = "SVM",
  engine = "kernlab",
  mode = "classification",
  kernel = "linear",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  gridNum = 5,
  iter = 15,
  metric = NULL,
  seed = 1234
)
}
\description{
SVM
}
\details{
SVM
}
