% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{SVMLinear}
\alias{SVMLinear}
\title{SVMLinear}
\usage{
SVMLinear(
  algo = "SVM",
  engine = "kernlab",
  mode = "classification",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  gridNum = 5,
  iter = 15,
  metric = NULL,
  seed = 1234
)
}
\description{
SVMLinear
}
\details{
SVMLinear
}
