% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{trainTestSplit}
\alias{trainTestSplit}
\title{Train-Test Split}
\usage{
trainTestSplit(data = data, target = NULL, prop, seed = "4814")
}
\arguments{
\item{data}{data}

\item{target}{target}

\item{prop}{prop}

\item{seed}{seed}
}
\description{
Train-Test Split
}
\details{
Data를 Train set과 Test set으로 분리합니다.
}
