% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{xgboost_phi}
\alias{xgboost_phi}
\title{XGBoost}
\usage{
xgboost_phi(engine = "xgboost", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
XGBoost
}
\details{
XGBoost
hyperparameters: mtry, min_n, tree_depth, loss_reduction, learn_rate, sample_size
}
