% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{clusteringVis}
\alias{clusteringVis}
\title{clusteringVis}
\usage{
clusteringVis(
  data = NULL,
  model = NULL,
  maxK = "15",
  nBoot = "100",
  selectOptimal = "silhouette",
  seedNum = "6471"
)
}
\arguments{
\item{data}{data}

\item{model}{model}

\item{maxK}{maxK}

\item{nStart}{nStart}
}
\description{
clusteringVis
}
\details{
Deprecated
}
