% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{clusteringVis}
\alias{clusteringVis}
\title{clusteringVis}
\usage{
clusteringVis(
  data = NULL,
  model = NULL,
  maxK = NULL,
  nStart = NULL,
  nBoot = 100,
  selectOptimal = NULL
)
}
\arguments{
\item{data}{data}

\item{model}{model}

\item{maxK}{maxK}

\item{nStart}{nStart}
}
\description{
clusteringVis
}
\details{
clusteringVis
}
