% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{decisionTree}
\alias{decisionTree}
\title{Decision Tree}
\usage{
decisionTree(
  algo = "Decision Tree",
  engine = "rpart",
  mode = "classification",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  gridNum = 5,
  iter = 10,
  metric = NULL,
  seed = 1234
)
}
\arguments{
\item{algo}{사용자가 임의로 지정할 알고리즘명 (default: "Decision Tree")}

\item{engine}{모델을 생성할 때 사용할 패키지 ("rpart" (default), "C50", "partykit")}

\item{mode}{분석 유형 ("classification" (default), "regression")}

\item{trainingData}{훈련데이터 셋}

\item{splitedData}{train-test 데이터 분할 정보를 포함하고 있는 전체 데이터 셋}

\item{formula}{모델링을 위한 수식}

\item{rec}{데이터, 전처리 정보를 포함한 recipe object}

\item{v}{v-fold cross validation을 진행 (default: 5, 각 fold 별로 30개 이상의 observations가 있어야 유효한 모델링 결과를 얻을 수 있습니다.)}

\item{metric}{모델의 성능을 평가할 기준지표 (classification: "roc_auc" (default), "accuracy" / regression: "rmse" (default), "rsq")}

\item{...}{hyperparameters의 범위에 대한 Min, Max, Levels 값에 해당하는 파라미터를 지정합니다.}
}
\description{
Decision Tree
}
\details{
의사결정나무 알고리즘 함수. 의사 결정 규칙 (Decision rule)을 나무 형태로 분류해 나가는 분석 기법을 말합니다.
hyperparameters:
tree_depth: 최종 예측값에 다다르기까지 몇 번 트리를 분할할지 설정합니다.
min_n: 트리를 분할하기 위해 필요한 관측값의 최소 개수를 설정합니다.
cost_complexity: 트리 분할을 위해 필요한 비용을 설정합니다. 0일 경우, 가능한 모든 분할이 수행됩니다.
}
