% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{decisionTree_phi}
\alias{decisionTree_phi}
\title{Decision Tree}
\usage{
decisionTree_phi(engine = "rpart", mode = "classification")
}
\arguments{
\item{engine}{engine}

\item{mode}{mode}
}
\description{
Decision Tree
}
\details{
의사결정나무 알고리즘 함수. 의사 결정 규칙 (Decision rule)을 나무 형태로 분류해 나가는 분석 기법을 말합니다.
hyperparameters: cost_complexity, tree_depth, min_n
}
