% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitBestModel}
\alias{fitBestModel}
\title{fitting in best model}
\usage{
fitBestModel(
  optResult,
  metric,
  model,
  formula,
  trainingData,
  splitedData,
  modelName
)
}
\arguments{
\item{optResult}{The result object of bayesOptCV}

\item{metric}{Baseline metric for evaluating model performance (classification: "roc_auc" (default), "accuracy" / regression: "rmse" (default), "rsq")}

\item{model}{The model object including the list of hyperparameters, engine and mode.}

\item{formula}{formula for modeling}

\item{trainingData}{The training data.}

\item{splitedData}{The whole dataset including the information of each fold}

\item{modelName}{The name of model defined by the algorithm and engine selected by user}
}
\description{
fitting in best model
}
\details{
Get the bayesOptCV function's return value and fit the model.
}
