% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{gridSearchCV}
\alias{gridSearchCV}
\title{Grid search with cross validation}
\usage{
gridSearchCV(
  rec = NULL,
  model = NULL,
  v = NULL,
  trainingData = NULL,
  parameterGrid = NULL,
  seed = NULL
)
}
\arguments{
\item{rec}{데이터, 전처리 정보를 포함한 recipe object}

\item{model}{hyperparameters, ngine, mode 정보가 포함된 model object}

\item{v}{v-fold cross validation을 진행 (default: 5, 각 fold 별로 30개 이상의 observations가 있어야 유효한 모델링 결과를 얻을 수 있습니다.)}

\item{trainingData}{훈련데이터 셋}

\item{seed}{seed값 설정}

\item{parameter_grid}{grid search를 수행할 때 각 hyperparameter의 값을 담은 object}
}
\description{
Grid search with cross validation
}
\details{
하이퍼파라미터를 탐색하는 Grid Search와 데이터 셋을 나누어 평가하는 cross validation을 함께 수행합니다.
}
