% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{linearRegression}
\alias{linearRegression}
\title{Linear Regression}
\usage{
linearRegression(
  algo = "Linear Regression",
  engine = "glmnet",
  mode = "regression",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  penaltyRangeMin = 0.001,
  penaltyRangeMax = 1,
  penaltyRangeLevels = 5,
  mixtureRangeMin = 0,
  mixtureRangeMax = 1,
  mixtureRangeLevels = 5,
  metric = "rmse"
)
}
\arguments{
\item{algo}{사용자가 임의로 지정할 알고리즘명 (default: "Linear Regression")}

\item{engine}{모델을 생성할 때 사용할 패키지 ("glmnet" (default), "lm", "glm", "stan")}

\item{mode}{분석 유형 ("regression" (default))}

\item{trainingData}{훈련데이터 셋}

\item{splitedData}{train-test 데이터 분할 정보를 포함하고 있는 전체 데이터 셋}

\item{formula}{모델링을 위한 수식}

\item{rec}{데이터, 전처리 정보를 포함한 recipe object}

\item{v}{v-fold cross validation을 진행 (default: 5, 각 fold 별로 30개 이상의 observations가 있어야 유효한 모델링 결과를 얻을 수 있습니다.)}

\item{metric}{모델의 성능을 평가할 기준지표 (classification: "roc_auc" (default), "accuracy" / regression: "rmse" (default), "rsq")}

\item{...}{hyperparameters의 범위에 대한 Min, Max, Levels 값에 해당하는 파라미터를 지정합니다.}
}
\description{
Linear Regression
}
\details{
선형 회귀 알고리즘 함수. 선형회귀는 다음과 같은 가정을 합니다. 1) target - features 간의 선형성, 2) features 간 작은 다중공선성,
3) 등분산성 가정, 4) 오차항의 정규분포, 5) 오차항 간 적은 상관성. 만약 데이터가 이 가정을 충족하지 않는 경우 성능이 저하될 수 있습니다.
hyperparameters: penalty, mixture
}
