% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{logisticRegression}
\alias{logisticRegression}
\title{Logistic Regression}
\usage{
logisticRegression(
  algo = "logistic Regression",
  engine = "glmnet",
  mode = "classification",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  gridNum = 5,
  iter = 10,
  metric = "roc_auc",
  seed = 1234
)
}
\arguments{
\item{algo}{A name of the algorithm which can be customized by user. (default: "logistic Regression")}

\item{engine}{The name of software that should be used to fit the model. (Option: "glmnet" (default))}

\item{mode}{The model type. It should be "classification" or "regression". (Option: "classification" (default))}

\item{trainingData}{A data frame for training}

\item{splitedData}{A data frame including metadata of split}

\item{formula}{formula for modeling}

\item{rec}{Recipe object containing preprocessing information for cross-validation}

\item{v}{Applying v-fold cross validation in modeling process (default: 5)}

\item{metric}{Metric to evaluate the performance (Option: "roc_auc" (default), "accuracy")}
}
\description{
Logistic regression is a algorithm for binary classification problems using a logistic function to model the probability of the positive class.

This function supports: binary classification

Assumption:
\itemize{
\item The modeling function using logistic regression
\item The dependent variable is binary.
\item The independent variables are linearly related to the log odds of the dependent variable.
\item There is little or no multicollinearity among independent variables.
\item There is little or no missing data.
\item The observations are independent of each other.
\item The error term has a constant variance and is normally distributed.
\item There is a large sample size.
}
}
\details{
Hyperparameters for tuning: penalty, mixture
}
