% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{logisticRegression}
\alias{logisticRegression}
\title{Logistic Regression}
\usage{
logisticRegression(
  algo = "Logistic Regression",
  engine = "glmnet",
  mode = "classification",
  trainingData = NULL,
  splitedData = NULL,
  formula = NULL,
  rec = NULL,
  v = 5,
  gridNum = 5,
  iter = 10,
  metric = "roc_auc",
  seed = 1234
)
}
\arguments{
\item{algo}{A name of the algorithm which can be customized by user (default: "Logistic Regression").}

\item{engine}{The name of software that should be used to fit the model (Option: "glmnet" (default)).}

\item{mode}{The model type. It should be "classification" or "regression" (Option: "classification" (default)).}

\item{trainingData}{The training data.}

\item{splitedData}{The whole dataset including the information of each fold}

\item{formula}{formula for modeling}

\item{rec}{Recipe object containing preprocessing information for cross-validation}

\item{v}{Applying v-fold cross validation in modeling process (default: 5)}

\item{gridNum}{Initial number of iterations to run before starting the optimization algorithm.}

\item{iter}{The maximum number of search iterations.}

\item{metric}{Metric to evaluate the performance (classification: "roc_auc" (default), "accuracy" / regression: "rmse" (default), "rsq").}

\item{seed}{Seed for reproducible results.}
}
\description{
The function for training user-defined Logistic regression model.

This function supports: binary classification
}
\details{
Hyperparameters for tuning: penalty, mixture
}
