% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{\%>\%}
\alias{\%>\%}
\alias{rocCurve}
\alias{confusionMatrix}
\alias{regressionPlot}
\alias{evalMetricsC}
\title{AUC-ROC Curve}
\usage{
rocCurve(modelsList, targetVar)

confusionMatrix(modelName, modelsList, targetVar)

regressionPlot(modelName, modelsList, targetVar)

evalMetricsC(modelsList, targetVar)
}
\arguments{
\item{modelsList}{ML 모델 리스트}

\item{targetVar}{타겟 변수}

\item{modelName}{모델명}
}
\description{
AUC-ROC Curve

Confusion matrix

Regression plot

Evaluation metrics for Classification
}
\details{
ML 모델 리스트로부터 AUC-ROC Curve를 생성합니다.

ML 모델 리스트 내 특정 모델에 대해 Confusion matrix를 생성합니다.

ML 모델 리스트 내 특정 모델에 대해 Regression plot를 생성합니다.

ML 모델 리스트로부터 Classification 모델들에 대한 Evaluation metrics를 생성합니다.
}
