% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepForCV}
\alias{prepForCV}
\title{Preprocessing for cross validation}
\usage{
prepForCV(
  data = NULL,
  formula = NULL,
  imputation = FALSE,
  normalization = FALSE,
  nomialImputationType = NULL,
  numericImputationType = NULL,
  normalizationType = NULL,
  seed = "4814"
)
}
\arguments{
\item{data}{data}

\item{formula}{formula}

\item{imputation}{imputation}

\item{normalization}{normalization}

\item{normalizationType}{normalizationType}

\item{seed}{seed}

\item{imputationType}{imputationType}
}
\description{
Preprocessing for cross validation
}
\details{
Preprocessing for cross validation
}
