% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepForCV}
\alias{prepForCV}
\title{Preprocessing for cross validation}
\usage{
prepForCV(
  data,
  formula,
  imputationType = "mean",
  normalizationType = "range",
  imputation = FALSE,
  normalization = FALSE,
  seed = "4814"
)
}
\arguments{
\item{data}{data}

\item{formula}{formula}

\item{imputationType}{imputationType}

\item{normalizationType}{normalizationType}

\item{imputation}{imputation}

\item{normalization}{normalization}

\item{seed}{seed}

\item{pcaThres}{pcaThres}

\item{pca}{pca}
}
\description{
Preprocessing for cross validation
}
\details{
Preprocessing for cross validation
}
