% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{trainTestSplit}
\alias{trainTestSplit}
\title{Train-Test Split}
\usage{
trainTestSplit(data = NULL, target = NULL, prop, seed = "4814")
}
\arguments{
\item{data}{전처리가 완료된 전체 data}

\item{target}{타겟 변수}

\item{prop}{전체 데이터 중 훈련 데이터로 사용할 비율}

\item{seed}{seed값 설정}
}
\description{
Train-Test Split
}
\details{
Data를 Train set과 Test set으로 분리합니다.
}
