% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{trainTestSplit}
\alias{trainTestSplit}
\title{Train-Test Split}
\usage{
trainTestSplit(data = data, target = NULL, prop)
}
\arguments{
\item{data}{data}

\item{prop}{prop}

\item{strata}{strata}
}
\description{
Train-Test Split
}
\details{
Train-Test Split
}
